primera versión funcional
Esta versión,  mide la temperatura y la imprime en el LCD
solo números enteros

+++ Código/Control_Temperatura/Control_Temperatura/termometro.h
@@ -6,4 +6,4 @@
  */ 
 #include <avr/io.h>
 void TERMOMETRO_init(void);
-int TERMOMETRO_get_temperatura(void);
+uint8_t TERMOMETRO_get_temperatura(void);
